Fix minor bug
