+++ package.json
@@ -16,6 +16,12 @@
 		"fix": "npx stylelint --fix \"**/*.{css,scss}\"",
 		"lintfix": "npx eslint --fix ."
 	},
+	"jest": {
+		"moduleNameMapper": {
+			"\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+			"\\.(css|less)$": "<rootDir>/__mocks__/fileMock.js"
+		}
+	},
 	"repository": {
 		"type": "git",
 		"url": "git+https://github.com/uimarshall/webpack-config.git"
